Move JsonItemConverterInterface to separate file
